requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
matplotlib
seaborn
requests
python-dotenv
scipy
streamlit

